% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyprod_deploy.R
\name{shinyprod_deploy}
\alias{shinyprod_deploy}
\title{Deploy workshop app}
\usage{
shinyprod_deploy(tag = "shinyprod")
}
\arguments{
\item{tag}{string for tag of the content. Default is "shinyprod"}
}
\description{
Deploy workshop app
}
